fixed tunneling

+++ src/Servers/Servers.jl
@@ -195,23 +195,24 @@ end
 
 function start(s::Server)
     @info "Starting:\n$s"
-    if s.local_port != 0
-        t = getfirst(x->occursin("ssh -f -N", x), split(read(pipeline(`ps aux` , stdout = `grep $(s.local_port)`), String), "\n"))
-        
-        if t !== nothing
-            run(`kill $(split(t)[2])`)
+    julia_cmd = """$(s.julia_exec) --startup-file=no -t auto -e "using DFControl; DFControl.Resource.run($(s.port))" &> ~/.julia/config/DFControl/errors.log"""
+    if s.domain != "localhost"
+        if s.local_port != 0
+            t = getfirst(x->occursin("ssh -f -N", x), split(read(pipeline(`ps aux` , stdout = `grep $(s.local_port)`), String), "\n"))
+            
+            if t !== nothing
+                run(`kill $(split(t)[2])`)
+            end
+            run(Cmd(`ssh -f -L $(s.local_port):localhost:$(s.port) $(ssh_string(s)) $julia_cmd`, detach=true))
+        else
+            run(Cmd(`ssh -f $(ssh_string(s)) $julia_cmd`, detach=true))
+            
         end
-        run(`ssh -f -N -L $(s.local_port):localhost:$(s.port) $(ssh_string(s))`)
+    else
+        run(Cmd(`$julia_cmd`, detach=true))
     end
+        
     #TODO: little hack here
-    proc = addprocs(s)[1]
-    hd = remotecall_fetch(homedir, proc)
-    
-    cmd = pipeline(Cmd(`$(s.julia_exec) --startup-file=no -t auto -e "using DFControl; DFControl.Resource.run($(s.port))"`;
-              detach = true), stderr = joinpath(hd, ".julia/config/DFControl/errors.log"))
-
-    p = remotecall(run, proc, cmd; wait = false)
-
     while !isalive(s)
         sleep(1)
     end
@@ -221,7 +222,7 @@ function start(s::Server)
     else
         @info "Daemon on Server $(s.name) started, listening on local port $(s.local_port)."
     end
-    return rmprocs(proc)
+    return 
 end
 
 function maybe_start_server(s::Server)
